lab3-worker-rest/requirements.txt
pymongo
python-dotenv
requests
structlog
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-requests